+++ src/CSharpLanguageServer/Handlers/Definition.fs
@@ -27,10 +27,12 @@ module Definition =
         match dynamicRegistration clientCapabilities with
         | false -> None
         | true ->
+            let registerOptions: DefinitionRegistrationOptions =
+                { DocumentSelector = Some defaultDocumentSelector }
             Some
                 { Id = Guid.NewGuid().ToString()
                   Method = "textDocument/definition"
-                  RegisterOptions = { DocumentSelector = Some defaultDocumentSelector } |> serialize |> Some }
+                  RegisterOptions = registerOptions |> serialize |> Some }
 
     let handle (context: ServerRequestContext) (p: TextDocumentPositionParams) : AsyncLspResult<GotoResult option> = async {
         match! context.FindSymbol' p.TextDocument.Uri p.Position with