+++ src/CSharpLanguageServer/Handlers/Hover.fs
@@ -28,10 +28,12 @@ module Hover =
         match dynamicRegistration clientCapabilities with
         | false -> None
         | true ->
+            let registerOptions: HoverRegistrationOptions =
+                { DocumentSelector = Some defaultDocumentSelector }
             Some
                 { Id = Guid.NewGuid().ToString()
                   Method = "textDocument/hover"
-                  RegisterOptions = { DocumentSelector = Some defaultDocumentSelector } |> serialize |> Some }
+                  RegisterOptions = registerOptions |> serialize |> Some }
 
     let handle (context: ServerRequestContext) (p: TextDocumentPositionParams) : AsyncLspResult<Hover option> = async {
         match! context.FindSymbol' p.TextDocument.Uri p.Position with