+++ src/CSharpLanguageServer/Handlers/References.fs
@@ -28,10 +28,11 @@ module References =
         match dynamicRegistration clientCapabilities with
         | false -> None
         | true ->
+            let registerOptions: ReferenceRegistrationOptions = { DocumentSelector = Some defaultDocumentSelector }
             Some
                 { Id = Guid.NewGuid().ToString()
                   Method = "textDocument/references"
-                  RegisterOptions = { DocumentSelector = Some defaultDocumentSelector } |> serialize |> Some }
+                  RegisterOptions = registerOptions |> serialize |> Some }
 
     let handle (context: ServerRequestContext) (p: ReferenceParams) : AsyncLspResult<Location[] option> = async {
         match! context.FindSymbol p.TextDocument.Uri p.Position with