+++ src/CSharpLanguageServer/Handlers/TypeDefinition.fs
@@ -29,10 +29,12 @@ module TypeDefinition =
         match dynamicRegistration clientCapabilities with
         | false -> None
         | true ->
+            let registerOptions: TypeDefinitionRegistrationOptions =
+                { DocumentSelector = Some defaultDocumentSelector }
             Some {
                 Id = Guid.NewGuid().ToString()
                 Method = "textDocument/typeDefinition"
-                RegisterOptions = { DocumentSelector = Some defaultDocumentSelector } |> serialize |> Some }
+                RegisterOptions = registerOptions |> serialize |> Some }
 
     let handle (context: ServerRequestContext) (p: TextDocumentPositionParams) : AsyncLspResult<GotoResult option> = async {
         match! context.FindSymbol' p.TextDocument.Uri p.Position with